requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# AI/ML
google-generativeai==0.3.2
openai==1.3.7

# Database
sqlalchemy==2.0.23
alembic==1.13.0

# Task Scheduling
apscheduler==3.10.4

# Email
aiosmtplib==3.0.1
email-validator==2.1.0

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Data Processing
pandas==2.1.4
numpy==1.26.2
openpyxl==3.1.2

# Document Generation
python-docx==1.1.0
reportlab==4.0.7
markdown==3.5.1

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.3

# Configuration
python-dotenv==1.0.0
pyyaml==6.0.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3
jinja2==3.1.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Code Quality
black==23.12.0
ruff==0.1.7
mypy==1.7.1

# Logging
loguru==0.7.2

# Web Scraping (optional)
beautifulsoup4==4.12.2
selenium==4.16.0

# Calendar
icalendar==5.0.11

# Cryptography
cryptography==41.0.7

# Redis (for caching/queue)
redis==5.0.1
